Modified arguments to start in SystemUserSpawner
The start method in SystemUserSpawner did not accept the same arguments as
the start mehtod in DockerSpawner. This made subclassing SystemUserSpawner
difficult as passing arguments was not possible.

This commit changes the arguments of start to match those of the
DockerSpawner method. It also modified the call to the super, passing in the
added arguments.

This commit does change the functionality a bit, as `working_dir` is no
longer forced to be `start.homedir` and can be overwritten.