Allow lookup of host homedir via `pwd`
Fixes issue jupyterhub/dockerspawner#299.
The home directory of the user on the host can be looked up via the
`pwd` database, which is useful when dealing with non-local users
(e.g., through LDAP authentication).

The lookup can be enabled by setting `host_homedir_format_string` to
an empty string or `None`.