Make sure create_object() creates the service task
Occasionally single-user task is not created immediately when create_service yields. Therefore, it is necessary to wait until it is created. Wait loop added to the end of create_object() method guarantees this. This problem is linked to Issue 330 and I think occurs more frequently if swarm includes nodes with different architectures (e.g. Arm64v8 and AMD64). The change also includes the patch by @gdbassett to the get_task() method to get task created by the service and additional states ("ready", "rejected") to check its state.